refactor: enhance WaitListForm error handling and API response validation
* Implemented zod schemas for API response validation.
* Improved error handling in WaitListForm to manage unknown responses and display appropriate error messages.
* Updated the import structure in types.ts to streamline API response types.
