+++ src/lib/api/schemas.ts
@@ -1,4 +1,5 @@
 import { z } from "zod";
+import type { ErrorCode } from "./constants";
 
 /**
  * Custom error messages in Portuguese
@@ -18,3 +19,38 @@ export const emailSchema = z
     required_error: errorMessages.email.required,
   })
   .email(errorMessages.email.invalid);
+
+/**
+ * Successful API Response Schema
+ */
+export const successResponseSchema = z.object({
+  success: z.literal(true),
+  message: z.string(),
+  code: z.never().optional(),
+  errors: z.never().optional(),
+});
+
+/**
+ * Error API Response Schema
+ */
+export const errorResponseSchema = z.object({
+  success: z.literal(false),
+  message: z.string(),
+  // The error code types are already being enforced on the API createResponse
+  // So they will not be enforced manually here, for the time being
+  code: z.string() as z.ZodType<ErrorCode>,
+  errors: z.array(
+    z.object({
+      path: z.string(),
+      message: z.string(),
+    }),
+  ),
+});
+
+/**
+ * API Response Schema
+ */
+export const apiResponseSchema = z.discriminatedUnion("success", [
+  successResponseSchema,
+  errorResponseSchema,
+]);
